[server] : refactor(UserToken): 불필요한 추상화 제거
serialize 되기 전에 User Entity 에서 UserToken으로만 변경되면 되기 때문에 불필요한 추상화를 제거했습니다.
